ft(renamed components and folders): refactored some code
