+++ src/App.js
@@ -4,7 +4,7 @@ import { Switch, Route } from "react-router-dom";
 import Header from "./components/Header/Header.component";
 import Homepage from "./Pages/Homepage/Homepage.component";
 import CategoriesPage from "./Pages/CategoriesPage/CategoriesPage.component";
-import CoursesPage from "./components/Courses/Courses.component";
+import CoursesOverview from "./components/Courses-Overview/Courses-Overview.component";
 import "./App.css";
 
 function App() {
@@ -14,7 +14,7 @@ function App() {
       <Switch>
         <Route exact path="/" component={Homepage} />
         <Route path="/categories" component={CategoriesPage} />
-        <Route path="/courses" component={CoursesPage} />
+        <Route path="/courses" component={CoursesOverview} />
       </Switch>
     </div>
   );
