+++ src/components/Category/Category.component.jsx
@@ -7,16 +7,13 @@ import {
   CategoryImg,
 } from "./Category.styles";
 
-const Category = ({ title, categoryImageUrl, courses }) => {
+const Category = ({ title, categoryImageUrl}) => {
   return (
     <CategoryContainer>
       <Card>
         <CategoryImg src={categoryImageUrl} />
         <Card.Body style={{ padding: "30px 35px" }}>
           <CategoryTitle>{title}</CategoryTitle>
-          <Card.Text style={{ color: "#fc3c64" }}>
-            {courses.length} <span>courses</span>{" "}
-          </Card.Text>
         </Card.Body>
       </Card>
     </CategoryContainer>
