+++ src/redux/categories/categories.data.js
@@ -1,0 +1,40 @@
+const CATEGORIES_DATA = [
+  {
+    id: 1,
+    title: "Programming",
+    categoryImageUrl: "https://i.ibb.co/fFQ0QVp/6.jpg",
+    linkUrl: "it-development",
+  },
+  {
+    id: 2,
+    title: "Graphic Design",
+    categoryImageUrl: "https://i.ibb.co/ZT5d6nb/classes-5.jpg",
+    linkUrl: "graphic-design",
+  },
+  {
+    id: 3,
+    title: "Marketing",
+    categoryImageUrl: "https://i.ibb.co/gVrxrxv/1.jpg",
+    linkUrl: "marketing",
+  },
+  {
+    id: 4,
+    title: "Web Development",
+    categoryImageUrl: "https://i.ibb.co/gVrxrxv/1.jpg",
+    linkUrl: "web-development",
+  },
+  {
+    id: 5,
+    title: "Language & Speech",
+    categoryImageUrl: "https://i.ibb.co/ZmpFhpP/2.jpg",
+    linkUrl: "language-speech",
+  },
+  {
+    id: 6,
+    title: "Business",
+    categoryImageUrl: "https://i.ibb.co/KX021w3/4.jpg",
+    linkUrl: "business",
+  },
+];
+
+export default CATEGORIES_DATA;
