+++ src/redux/courses/courses.data.js
@@ -1,0 +1,184 @@
+const COURSES_DATA = [
+  {
+    id: 101,
+    title: "Javascript for beginners",
+    category: "Programming",
+    courseImageUrl: "https://i.ibb.co/j83ZG5Q/f-2.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 102,
+    title: "Developing Mobile Apps",
+    category: "Programming",
+    courseImageUrl: "https://i.ibb.co/jJLx4Pt/classes-1.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 103,
+    title: "Python for Beginners",
+    category: "Programming",
+    courseImageUrl: "https://i.ibb.co/gVrxrxv/1.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 104,
+    title: "Programming for Kids",
+    category: "Programming",
+    courseImageUrl: "https://i.ibb.co/1q0ry2h/course-5.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 105,
+    title: "Adobe Photoshop",
+    category: "Graphic Design",
+    courseImageUrl: "https://i.ibb.co/Jp2wC7s/classes-2.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 106,
+    title: "Creating Banners",
+    category: "Graphic Design",
+    courseImageUrl: "https://i.ibb.co/LZ4yp0W/1.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 107,
+    title: "Adobe Sketch XD",
+    category: "Graphic Design",
+    courseImageUrl: "https://i.ibb.co/QQ8RGSZ/3.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 108,
+    title: "Photography",
+    category: "Graphic Design",
+    courseImageUrl: "https://i.ibb.co/1Qy8Jzb/3.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 109,
+    title: "Social Media",
+    category: "Marketing",
+    courseImageUrl: "https://i.ibb.co/9GDcHh3/classes-8.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 110,
+    title: "Network Marketing",
+    category: "Marketing",
+    courseImageUrl: "https://i.ibb.co/VxHHB5Y/5.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 112,
+    title: "Social Media Content Management Social Media Content Management",
+    category: "Marketing",
+    courseImageUrl: "https://i.ibb.co/nMDBn1g/4.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/nnTwrx1/3.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 113,
+    title: "Frontend Development",
+    category: "Web Development",
+    courseImageUrl: "https://i.ibb.co/MV41pkN/2.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/4FWHH83/5.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 116,
+    title: "Agile & Scrum Methedologies",
+    category: "Web Development",
+    courseImageUrl: "https://i.ibb.co/qg4drn2/course-11.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/qkQCp5t/6.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 115,
+    title: "Learn Spanish",
+    category: "Language & Speech",
+    courseImageUrl: "https://i.ibb.co/WPPbVyR/course-2.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/fqLfCpS/7.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 155,
+    title: "Learn English",
+    category: "Language & Speech",
+    courseImageUrl: "https://i.ibb.co/FsTCHnF/course-9.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/6NBFG8M/4.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 111,
+    title: "Learn Poetry",
+    category: "Marketing",
+    courseImageUrl: "https://i.ibb.co/K64nHgW/course-8.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/Vw7V2hz/2.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 16,
+    title: "Business Analytics",
+    category: "Business",
+    courseImageUrl: "https://i.ibb.co/KbNNHQJ/course-4.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/6YPxGT3/1.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+  {
+    id: 114,
+    title: "Team Organisation",
+    category: "Business",
+    courseImageUrl: "https://i.ibb.co/bPhV6VH/classes-7.jpg",
+    numberOfStudents: 20,
+    authorImageUrl: "https://i.ibb.co/chGDkz0/8.jpg",
+    authorName: "Richard Hendricks",
+    ratings: 4.5,
+  },
+];
+
+export default COURSES_DATA;
