+++ src/redux/courses/courses.reducer.jsx
@@ -1,0 +1,14 @@
+import COURSES_DATA from "./courses.data";
+
+const INITIAL_STATE = {
+  courses: COURSES_DATA,
+};
+
+const coursesReducer = (state = INITIAL_STATE, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
+
+export default coursesReducer;
