+++ src/redux/courses/courses.selectors.jsx
@@ -1,0 +1,8 @@
+import { createSelector } from "reselect";
+
+const selectCourses = (state) => state.courses;
+
+export const selectCoursesSections = createSelector(
+  [selectCourses],
+  (courses) => courses.courses
+);
