+++ src/redux/root-reducer.js
@@ -1,7 +1,9 @@
-import {combineReducers} from 'redux';
+import { combineReducers } from "redux";
 
-import categoriesReducer from './categories/categories.reducer';
+import categoriesReducer from "./categories/categories.reducer";
+import coursesReducer from "./courses/courses.reducer";
 
 export default combineReducers({
-    categories: categoriesReducer
-})
+  categories: categoriesReducer,
+  courses: coursesReducer,
+});
